feat: remove separate classes for bid/ask events
