+++ packages/indexer/src/jobs/websocket-events/events/index.ts
@@ -1,4 +1,2 @@
 import "@/jobs/websocket-events/events/new-top-bid-websocket-event";
 import "@/jobs/websocket-events/events/new-activity-websocket-event";
-import "@/jobs/websocket-events/events/ask-websocket-event";
-import "@/jobs/websocket-events/events/bid-websocket-event";
